move python code execution logic out of builtins
